refactor: Participants 인터페이스 -> 클래스로 변경
