+++ src/main/java/racingGame/RacingGameApplication.java
@@ -5,7 +5,6 @@ import racingGame.game.GameResult;
 import racingGame.game.RacingGame;
 import racingGame.game.RacingGameRule;
 import racingGame.participant.Participants;
-import racingGame.participant.ParticipantsV2;
 import racingGame.view.InputView;
 import racingGame.view.ResultView;
 
@@ -15,7 +14,7 @@ public class RacingGameApplication {
         InputView inputView = new InputView();
         inputView.execute();
 
-        Participants participants = new ParticipantsV2(inputView.getParticipates());
+        Participants participants = new Participants(inputView.getParticipates());
         int gameCount = inputView.getGameCount();
         RacingGameRule gameRule = new CarForwardRule();
         RacingGame racingGame = new RacingGame(gameRule);
